Standardize playback control style
