+++ App/Views/Viewers/Podcast Episode/PodcastEpisodeView.swift
@@ -94,7 +94,7 @@ struct PodcastEpisodeView: View {
                         }
                         .aspectRatio(contentMode: .fit)
                         .frame(maxWidth: 300, maxHeight: 300)
-                        .clipShape(RoundedRectangle(cornerRadius: 16))
+                        .clipShape(.rect(cornerRadius: 16))
                         .overlay(
                             RoundedRectangle(cornerRadius: 16)
                                 .stroke(.quaternary, lineWidth: 0.5)
@@ -106,7 +106,7 @@ struct PodcastEpisodeView: View {
                             .resizable()
                             .aspectRatio(contentMode: .fit)
                             .frame(maxWidth: 300, maxHeight: 300)
-                            .clipShape(RoundedRectangle(cornerRadius: 16))
+                            .clipShape(.rect(cornerRadius: 16))
                             .overlay(
                                 RoundedRectangle(cornerRadius: 16)
                                     .stroke(.quaternary, lineWidth: 0.5)
@@ -145,10 +145,9 @@ struct PodcastEpisodeView: View {
                             }
                         }
                     }
-                    .padding(.horizontal)
 
                     if isThisEpisode {
-                        VStack(spacing: 12) {
+                        VStack(spacing: 16) {
                             SeekBarView(
                                 currentTime: Binding(
                                     get: { audioPlayer.currentTime },
@@ -158,35 +157,14 @@ struct PodcastEpisodeView: View {
                                 onSeek: { audioPlayer.seek(to: $0) }
                             )
 
-                            HStack {
-                                transcriptToggle
-                                    .frame(maxWidth: .infinity, alignment: .leading)
-
-                                HStack(spacing: 40) {
-                                    Button { audioPlayer.skipBackward() } label: {
-                                        Image(systemName: "gobackward.15")
-                                            .font(.title2)
-                                    }
-
-                                    Button { audioPlayer.togglePlayPause() } label: {
-                                        Image(systemName: audioPlayer.isPlaying
-                                              ? "pause.circle.fill"
-                                              : "play.circle.fill")
-                                            .font(.system(size: 72))
-                                    }
-
-                                    Button { audioPlayer.skipForward() } label: {
-                                        Image(systemName: "goforward.30")
-                                            .font(.title2)
-                                    }
-                                }
-
-                                playbackSpeedMenu
-                                    .frame(maxWidth: .infinity, alignment: .trailing)
-                            }
-                            .foregroundStyle(.primary)
+                            PodcastPlayerControls(
+                                audioPlayer: audioPlayer,
+                                hasTranscript: transcript != nil,
+                                showingTranscript: $showingTranscript,
+                                playbackSpeed: $playbackSpeed,
+                                playbackSpeedPresets: playbackSpeedPresets
+                            )
                         }
-                        .padding(.horizontal)
                     } else {
                         HStack(spacing: 12) {
                             Button {
@@ -252,9 +230,8 @@ struct PodcastEpisodeView: View {
                     .animation(.smooth.speed(2.0), value: showingTranslation)
                     .animation(.smooth.speed(2.0), value: showingTranscript)
                     .animation(.smooth.speed(2.0), value: translatedText)
-                    .padding(.horizontal)
                 }
-                .padding(.vertical)
+                .padding()
             }
             .onScrollPhaseChange { _, newPhase in
                 if showingTranscript, isTranscriptAutoScrolling,
